通讯录好友

+++ GameGroup/PublicClass/Class/GameCommon.m
@@ -167,7 +167,7 @@ static GameCommon *my_gameCommon = NULL;
 +(NSString *)getCurrentTime{
     
     NSTimeInterval nowTime = [[NSDate date] timeIntervalSince1970];
-    return [NSString stringWithFormat:@"%f",nowTime];
+    return [NSString stringWithFormat:@"%ld",(long)nowTime];
     
 }
 #pragma mark -周
